requirements.txt
torch>=1.12.0
numpy
torchvision>=0.13.0
requests
Pillow
scipy>=1.7.3
tqdm
matplotlib